+++ packages/evo-objects/tests/typed/__init__.py
@@ -9,26 +9,4 @@
 #  See the License for the specific language governing permissions and
 #  limitations under the License.
 
-from .regular_grid import (
-    Regular3DGrid,
-    Regular3DGridData,
-)
-from .regular_masked_grid import RegularMasked3DGrid, RegularMasked3DGridData
-from .tensor_grid import Tensor3DGrid, Tensor3DGridData
-from .types import BoundingBox, CoordinateReferenceSystem, EpsgCode, Point3, Rotation, Size3d, Size3i
-
-__all__ = [
-    "BoundingBox",
-    "CoordinateReferenceSystem",
-    "EpsgCode",
-    "Point3",
-    "Regular3DGrid",
-    "Regular3DGridData",
-    "RegularMasked3DGrid",
-    "RegularMasked3DGridData",
-    "Rotation",
-    "Size3d",
-    "Size3i",
-    "Tensor3DGrid",
-    "Tensor3DGridData",
-]
+"""Tests for typed geoscience objects."""
