+++ packages/evo-objects/tests/typed/test_regular_grid.py
@@ -1,0 +1,218 @@
+#  Copyright © 2025 Bentley Systems, Incorporated
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#      http://www.apache.org/licenses/LICENSE-2.0
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License.
+
+from __future__ import annotations
+
+import contextlib
+import dataclasses
+import uuid
+from unittest.mock import patch
+
+import numpy as np
+import pandas as pd
+
+from evo.common import Environment, EvoContext
+from evo.common.test_tools import BASE_URL, ORG, WORKSPACE_ID, TestWithConnector
+from evo.objects import ObjectReference
+from evo.objects.typed import Point3, Regular3DGrid, Regular3DGridData, Rotation, Size3d, Size3i
+from evo.objects.typed.dataset import DataLoaderError
+from evo.objects.typed.exceptions import ObjectValidationError
+
+from .helpers import MockClient
+
+
+class TestRegularGrid(TestWithConnector):
+    def setUp(self) -> None:
+        TestWithConnector.setUp(self)
+        self.environment = Environment(hub_url=BASE_URL, org_id=ORG.id, workspace_id=WORKSPACE_ID)
+        self.context = EvoContext.from_environment(
+            environment=self.environment,
+            connector=self.connector,
+        )
+
+    @contextlib.contextmanager
+    def _mock_geoscience_objects(self):
+        mock_client = MockClient(self.environment)
+        with (
+            patch("evo.objects.typed.dataset.get_data_client", lambda _: mock_client),
+            patch("evo.objects.typed.base.create_geoscience_object", mock_client.create_geoscience_object),
+            patch("evo.objects.typed.base.replace_geoscience_object", mock_client.replace_geoscience_object),
+            patch("evo.objects.typed.base.download_geoscience_object", mock_client.from_reference),
+        ):
+            yield mock_client
+
+    example_grid = Regular3DGridData(
+        name="Test Grid",
+        origin=Point3(0, 0, 0),
+        size=Size3i(10, 10, 5),
+        cell_size=Size3d(2.5, 5, 5),
+        cell_data=pd.DataFrame(
+            {
+                "value": np.random.rand(10 * 10 * 5),
+                "cat": pd.Categorical(np.random.choice(range(4), size=10 * 10 * 5), ["a", "b", "c", "d"]),
+            }
+        ),
+        vertex_data=pd.DataFrame(
+            {
+                "elevation": np.random.rand(11 * 11 * 6),
+            }
+        ),
+        rotation=Rotation(90, 0, 0),
+    )
+
+    async def test_create(self):
+        with self._mock_geoscience_objects():
+            result = await Regular3DGrid.create(evo_context=self.context, data=self.example_grid)
+        self.assertEqual(result.name, "Test Grid")
+        self.assertEqual(result.origin, Point3(0, 0, 0))
+        self.assertEqual(result.size, Size3i(10, 10, 5))
+        self.assertEqual(result.cell_size, Size3d(2.5, 5, 5))
+        self.assertEqual(result.rotation, Rotation(90, 0, 0))
+
+        cell_df = await result.cells.as_dataframe()
+        pd.testing.assert_frame_equal(cell_df, self.example_grid.cell_data)
+        vertices_df = await result.vertices.as_dataframe()
+        pd.testing.assert_frame_equal(vertices_df, self.example_grid.vertex_data)
+
+    async def test_replace(self):
+        data = dataclasses.replace(self.example_grid, vertex_data=None)
+        with self._mock_geoscience_objects():
+            result = await Regular3DGrid.replace(
+                evo_context=self.context,
+                reference=ObjectReference.new(
+                    environment=self.context.get_environment(),
+                    object_id=uuid.uuid4(),
+                ),
+                data=data,
+            )
+        self.assertEqual(result.name, "Test Grid")
+        self.assertEqual(result.origin, Point3(0, 0, 0))
+        self.assertEqual(result.size, Size3i(10, 10, 5))
+        self.assertEqual(result.cell_size, Size3d(2.5, 5, 5))
+        self.assertEqual(result.rotation, Rotation(90, 0, 0))
+
+        cell_df = await result.cells.as_dataframe()
+        pd.testing.assert_frame_equal(cell_df, data.cell_data)
+        vertices_df = await result.vertices.as_dataframe()
+        self.assertEqual(vertices_df.shape[0], 0)  # No vertex data provided
+
+    async def test_from_reference(self):
+        with self._mock_geoscience_objects():
+            original = await Regular3DGrid.create(evo_context=self.context, data=self.example_grid)
+
+            result = await Regular3DGrid.from_reference(evo_context=self.context, reference=original.metadata.url)
+            self.assertEqual(result.name, "Test Grid")
+            self.assertEqual(result.origin, Point3(0, 0, 0))
+            self.assertEqual(result.size, Size3i(10, 10, 5))
+            self.assertEqual(result.cell_size, Size3d(2.5, 5, 5))
+            self.assertEqual(result.rotation, Rotation(90, 0, 0))
+
+            cell_df = await result.cells.as_dataframe()
+            pd.testing.assert_frame_equal(cell_df, self.example_grid.cell_data)
+            vertices_df = await result.vertices.as_dataframe()
+            pd.testing.assert_frame_equal(vertices_df, self.example_grid.vertex_data)
+
+    async def test_update(self):
+        with self._mock_geoscience_objects():
+            obj = await Regular3DGrid.create(evo_context=self.context, data=self.example_grid)
+
+            self.assertEqual(obj.metadata.version_id, "1")
+            obj.name = "Updated Grid"
+            obj.origin = Point3(1, 1, 1)
+            await obj.cells.set_dataframe(
+                pd.DataFrame(
+                    {
+                        "value": np.ones(10 * 10 * 5),
+                    }
+                )
+            )
+
+            with self.assertRaises(DataLoaderError):
+                await obj.cells.as_dataframe()
+
+            await obj.update()
+
+            self.assertEqual(obj.name, "Updated Grid")
+            self.assertEqual(obj.origin, Point3(1, 1, 1))
+            self.assertEqual(obj.metadata.version_id, "2")
+
+            cell_df = await obj.cells.as_dataframe()
+            pd.testing.assert_frame_equal(
+                cell_df,
+                pd.DataFrame(
+                    {
+                        "value": np.ones(10 * 10 * 5),
+                    }
+                ),
+            )
+
+    async def test_size_check(self):
+        with self.assertRaises(ValueError):
+            dataclasses.replace(self.example_grid, size=Size3i(15, 10, 6))
+
+        with self._mock_geoscience_objects():
+            obj = await Regular3DGrid.create(evo_context=self.context, data=self.example_grid)
+            with self.assertRaises(ObjectValidationError):
+                await obj.cells.set_dataframe(
+                    pd.DataFrame(
+                        {
+                            "value": np.random.rand(11 * 10 * 5),
+                        }
+                    )
+                )
+
+            obj.size = Size3i(5, 10, 6)
+            with self.assertRaises(ObjectValidationError):
+                obj.validate()
+
+    async def test_size_check_data(self):
+        with self.assertRaises(ObjectValidationError):
+            Regular3DGridData(
+                name="Test Grid",
+                origin=Point3(0, 0, 0),
+                size=Size3i(10, 10, 5),
+                cell_size=Size3d(2.5, 5, 5),
+                cell_data=pd.DataFrame(
+                    {
+                        "value": np.random.rand(12 * 10 * 5),
+                    }
+                ),
+                rotation=Rotation(90, 0, 0),
+            )
+
+    async def test_bounding_box(self):
+        with self._mock_geoscience_objects() as mock_client:
+            obj = await Regular3DGrid.create(evo_context=self.context, data=self.example_grid)
+
+            bbox = obj.bounding_box
+            self.assertAlmostEqual(bbox.min_x, 0.0)
+            self.assertAlmostEqual(bbox.min_y, -25.0)
+            self.assertAlmostEqual(bbox.min_z, 0.0)
+            self.assertAlmostEqual(bbox.max_x, 50.0)
+            self.assertAlmostEqual(bbox.max_y, 0.0)
+            self.assertAlmostEqual(bbox.max_z, 25.0)
+
+            bbox = mock_client.objects[str(obj.metadata.url.object_id)]["bounding_box"]
+            self.assertAlmostEqual(bbox["min_x"], 0.0)
+            self.assertAlmostEqual(bbox["min_y"], -25.0)
+            self.assertAlmostEqual(bbox["min_z"], 0.0)
+            self.assertAlmostEqual(bbox["max_x"], 50.0)
+            self.assertAlmostEqual(bbox["max_y"], 0.0)
+            self.assertAlmostEqual(bbox["max_z"], 25.0)
+
+            obj.origin = Point3(1, 1, 1)
+            bbox = obj.bounding_box
+            self.assertAlmostEqual(bbox.min_x, 1.0)
+            self.assertAlmostEqual(bbox.min_y, -24.0)
+            self.assertAlmostEqual(bbox.min_z, 1.0)
+            self.assertAlmostEqual(bbox.max_x, 51.0)
+            self.assertAlmostEqual(bbox.max_y, 1.0)
+            self.assertAlmostEqual(bbox.max_z, 26.0)
